refactor: extract contract error types into errors.rs
- Create errors.rs with unified ContractError enum
- Consolidate all error definitions
- Update error.rs to re-export from errors module
- Update lib.rs to expose errors module
- Maintain backward compatibility with StellarSaveError
