+++ contracts/stellar-save/src/lib.rs
@@ -25,6 +25,7 @@ pub mod contribution;
 pub mod cycle_advancement;
 pub mod deadline;
 pub mod error;
+pub mod errors;
 pub mod events;
 pub mod group;
 pub mod migration;
@@ -61,6 +62,7 @@ mod tests;
 pub use contribution::{ContributionPage, ContributionRecord};
 use core::cmp;
 pub use error::{ContractResult, ErrorCategory, StellarSaveError};
+pub use errors::{ContractError, ErrorRecoveryStrategy};
 pub use events::EventEmitter;
 pub use events::*;
 pub use group::{Group, GroupStatus};
